Add missing fields to account verification
This allows the Mona iOS app to successfully sign in.

I wasn't sure what a suitable value for `last_status_at` would be so I
just copied `published` since it's the right format.

This also just sets the `bot` field to false regardless of the user's
configuration, which probably isn't ideal.