Fix custom emojis not appearing
This fixes emojis not appearing if source instance hasn't set the mediaType field.
We use xs_mime_by_ext in order to guess whether the file is SVG or not when mediaType is missing.
This commit also changes the behaviour to not replace :shortnames: with nothing when the emoji replacement failed.